Draft implementation of hook processing in tasks
